refactor (espt): change detail to inside espt package
