+++ internal/models/espt/detaildto.go
@@ -1,13 +1,13 @@
 package espt
 
 import (
-	"github.com/bapenda-kota-malang/apin-backend/internal/models/detailesptair"
-	"github.com/bapenda-kota-malang/apin-backend/internal/models/detailespthiburan"
-	"github.com/bapenda-kota-malang/apin-backend/internal/models/detailespthotel"
-	"github.com/bapenda-kota-malang/apin-backend/internal/models/detailesptparkir"
-	"github.com/bapenda-kota-malang/apin-backend/internal/models/detailesptppjnonpln"
-	"github.com/bapenda-kota-malang/apin-backend/internal/models/detailesptppjpln"
-	"github.com/bapenda-kota-malang/apin-backend/internal/models/detailesptresto"
+	"github.com/bapenda-kota-malang/apin-backend/internal/models/espt/detailesptair"
+	"github.com/bapenda-kota-malang/apin-backend/internal/models/espt/detailespthiburan"
+	"github.com/bapenda-kota-malang/apin-backend/internal/models/espt/detailespthotel"
+	"github.com/bapenda-kota-malang/apin-backend/internal/models/espt/detailesptparkir"
+	"github.com/bapenda-kota-malang/apin-backend/internal/models/espt/detailesptppjnonpln"
+	"github.com/bapenda-kota-malang/apin-backend/internal/models/espt/detailesptppjpln"
+	"github.com/bapenda-kota-malang/apin-backend/internal/models/espt/detailesptresto"
 )
 
 type CreateDetailAirDto struct {
