+++ internal/models/espt/detailespthiburan/detailespthiburan.go
@@ -1,7 +1,7 @@
 package detailespthiburan
 
 import (
-	"github.com/bapenda-kota-malang/apin-backend/internal/models/detailesptair"
+	"github.com/bapenda-kota-malang/apin-backend/internal/models/espt/detailesptair"
 	"github.com/bapenda-kota-malang/apin-backend/pkg/gormhelper"
 	"github.com/lib/pq"
 )
