+++ internal/models/espt/detailesptppjnonpln/detailesptppjnonpln.go
@@ -1,7 +1,7 @@
 package detailesptppjnonpln
 
 import (
-	"github.com/bapenda-kota-malang/apin-backend/internal/models/detailesptair"
+	"github.com/bapenda-kota-malang/apin-backend/internal/models/espt/detailesptair"
 	"github.com/bapenda-kota-malang/apin-backend/pkg/gormhelper"
 )
