+++ internal/models/espt/detailesptppjpln/detailesptppjpln.go
@@ -1,7 +1,7 @@
 package detailesptppjpln
 
 import (
-	"github.com/bapenda-kota-malang/apin-backend/internal/models/detailesptair"
+	"github.com/bapenda-kota-malang/apin-backend/internal/models/espt/detailesptair"
 	"github.com/bapenda-kota-malang/apin-backend/internal/models/jenisppj"
 	"github.com/bapenda-kota-malang/apin-backend/pkg/gormhelper"
 )
