+++ internal/services/espt/detailesptair/detailesptair.go
@@ -1,4 +1,4 @@
-package detailespthiburan
+package detailesptair
 
 import (
 	"errors"
@@ -11,23 +11,18 @@ import (
 	rp "github.com/bapenda-kota-malang/apin-backend/pkg/apicore/responses"
 	sh "github.com/bapenda-kota-malang/apin-backend/pkg/servicehelper"
 
-	m "github.com/bapenda-kota-malang/apin-backend/internal/models/detailespthiburan"
+	m "github.com/bapenda-kota-malang/apin-backend/internal/models/espt/detailesptair"
 
 	t "github.com/bapenda-kota-malang/apin-backend/pkg/apicore/types"
 )
 
-const source = "detailespthiburan"
+const source = "detailesptair"
 
 func Create(input m.CreateDto, tx *gorm.DB) (any, error) {
 	if tx == nil {
 		tx = a.DB
 	}
-
-	if input.KarcisBebas && input.JumlahKarcisBebas == nil {
-		return nil, errors.New("jumlah karcis bebas kosong")
-	}
-
-	var data m.DetailEsptHiburan
+	var data m.DetailEsptAir
 
 	//  copy input (payload) ke struct data jika tidak ada akan error
 	if err := sc.Copy(&data, &input); err != nil {
@@ -55,7 +50,7 @@ func Update(id int, input m.UpdateDto, tx *gorm.DB) (any, error) {
 	if tx == nil {
 		tx = a.DB
 	}
-	var data m.DetailEsptHiburan
+	var data m.DetailEsptAir
 
 	// validate data exist and copy input (payload) ke struct data jika tidak ada akan error
 	if id != 0 {
@@ -88,7 +83,7 @@ func Delete(id int, tx *gorm.DB) (any, error) {
 	if tx == nil {
 		tx = a.DB
 	}
-	var data *m.DetailEsptHiburan
+	var data *m.DetailEsptAir
 	result := tx.First(&data, id)
 	if result.RowsAffected == 0 {
 		return nil, errors.New("data tidak dapat ditemukan")
