+++ internal/services/espt/detailespthiburan/detailespthiburan.go
@@ -1,4 +1,4 @@
-package detailesptresto
+package detailespthiburan
 
 import (
 	"errors"
@@ -11,18 +11,23 @@ import (
 	rp "github.com/bapenda-kota-malang/apin-backend/pkg/apicore/responses"
 	sh "github.com/bapenda-kota-malang/apin-backend/pkg/servicehelper"
 
-	m "github.com/bapenda-kota-malang/apin-backend/internal/models/detailesptresto"
+	m "github.com/bapenda-kota-malang/apin-backend/internal/models/espt/detailespthiburan"
 
 	t "github.com/bapenda-kota-malang/apin-backend/pkg/apicore/types"
 )
 
-const source = "detailesptresto"
+const source = "detailespthiburan"
 
 func Create(input m.CreateDto, tx *gorm.DB) (any, error) {
 	if tx == nil {
 		tx = a.DB
 	}
-	var data m.DetailEsptResto
+
+	if input.KarcisBebas && input.JumlahKarcisBebas == nil {
+		return nil, errors.New("jumlah karcis bebas kosong")
+	}
+
+	var data m.DetailEsptHiburan
 
 	//  copy input (payload) ke struct data jika tidak ada akan error
 	if err := sc.Copy(&data, &input); err != nil {
@@ -50,7 +55,7 @@ func Update(id int, input m.UpdateDto, tx *gorm.DB) (any, error) {
 	if tx == nil {
 		tx = a.DB
 	}
-	var data m.DetailEsptResto
+	var data m.DetailEsptHiburan
 
 	// validate data exist and copy input (payload) ke struct data jika tidak ada akan error
 	if id != 0 {
@@ -83,7 +88,7 @@ func Delete(id int, tx *gorm.DB) (any, error) {
 	if tx == nil {
 		tx = a.DB
 	}
-	var data *m.DetailEsptResto
+	var data *m.DetailEsptHiburan
 	result := tx.First(&data, id)
 	if result.RowsAffected == 0 {
 		return nil, errors.New("data tidak dapat ditemukan")
